refactor: Change Money amount type from Long to Int and update related logic
- Replace Long with Int in Money class
- Update validation, calculations, and affected test assertions
- Adjust usage across LottoStoreTest, MoneyTest, and controller logic

+++ src/main/kotlin/lotto/Lottos.kt
@@ -20,7 +20,7 @@ class Lottos(private val values: List<Lotto>) {
     }
 
     private fun calculateTotalPrize(winningLotto: WinningLotto): Money {
-        val totalPrize: Long = matchAll(winningLotto)
+        val totalPrize: Int = matchAll(winningLotto)
             .entries
             .sumOf { (rank, count) -> rank.prize * count }
